webkit/data/layout_tests/chrome/fast/js/date-proto-generic-invocation-expected.txt

This test verifies that the functions of the Date prototype object are not generic, as documented in ECMA-262 rev3 section 15.9.5 Properties of the Date Prototype Object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS testFunctionName() threw exception TypeError.
PASS successfullyParsed is true

TEST COMPLETE
